# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Portfolio API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Portfolio projects
  - name: certificates
    description: Certificates
  - name: experiences
    description: Work & education experiences
  - name: contacts
    description: Contact form messages
  - name: skills
    description: Skills and technologies
  - name: admin
    description: Admin operations
  - name: stats
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: false
        - in: query
          name: featured
          schema:
            type: boolean
          required: false
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List all certificates
      responses:
        "200":
          description: List of certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certificate"
    post:
      operationId: createCertificate
      tags: [certificates]
      summary: Create a certificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"

  /certificates/{id}:
    patch:
      operationId: updateCertificate
      tags: [certificates]
      summary: Update a certificate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
    delete:
      operationId: deleteCertificate
      tags: [certificates]
      summary: Delete a certificate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /experiences:
    get:
      operationId: listExperiences
      tags: [experiences]
      summary: List all experiences
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [work, education]
          required: false
      responses:
        "200":
          description: List of experiences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experience"
    post:
      operationId: createExperience
      tags: [experiences]
      summary: Create an experience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"

  /experiences/{id}:
    patch:
      operationId: updateExperience
      tags: [experiences]
      summary: Update an experience
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
    delete:
      operationId: deleteExperience
      tags: [experiences]
      summary: Delete an experience
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List all skills
      parameters:
        - in: query
          name: category
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Create a skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"

  /skills/{id}:
    patch:
      operationId: updateSkill
      tags: [skills]
      summary: Update a skill
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete a skill
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contact messages (admin)
      responses:
        "200":
          description: List of contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

  /contacts/{id}/read:
    patch:
      operationId: markContactRead
      tags: [contacts]
      summary: Mark a contact message as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Unauthorized

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Admin session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get dashboard statistics summary
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"

  /stats/skills-by-category:
    get:
      operationId: getSkillsByCategory
      tags: [stats]
      summary: Get skills grouped by category with counts
      responses:
        "200":
          description: Skills by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkillCategoryCount"

  /projects/featured:
    get:
      operationId: listFeaturedProjects
      tags: [projects]
      summary: Get featured projects for homepage
      responses:
        "200":
          description: Featured projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Project:
      type: object
      required: [id, title, description, category, tags, featured, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        longDescription:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        liveUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        createdAt:
          type: string

    ProjectInput:
      type: object
      required: [title, description, category, tags]
      properties:
        title:
          type: string
        description:
          type: string
        longDescription:
          type: string
        imageUrl:
          type: string
        liveUrl:
          type: string
        githubUrl:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean

    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        longDescription:
          type: string
        imageUrl:
          type: string
        liveUrl:
          type: string
        githubUrl:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean

    Certificate:
      type: object
      required: [id, title, issuer, issueDate, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        issuer:
          type: string
        issueDate:
          type: string
        expiryDate:
          type: ["string", "null"]
        credentialUrl:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    CertificateInput:
      type: object
      required: [title, issuer, issueDate]
      properties:
        title:
          type: string
        issuer:
          type: string
        issueDate:
          type: string
        expiryDate:
          type: string
        credentialUrl:
          type: string
        imageUrl:
          type: string

    CertificateUpdate:
      type: object
      properties:
        title:
          type: string
        issuer:
          type: string
        issueDate:
          type: string
        expiryDate:
          type: string
        credentialUrl:
          type: string
        imageUrl:
          type: string

    Experience:
      type: object
      required: [id, title, organization, type, startDate, current, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        organization:
          type: string
        type:
          type: string
        description:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        current:
          type: boolean
        location:
          type: ["string", "null"]
        createdAt:
          type: string

    ExperienceInput:
      type: object
      required: [title, organization, type, startDate]
      properties:
        title:
          type: string
        organization:
          type: string
        type:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        current:
          type: boolean
        location:
          type: string

    ExperienceUpdate:
      type: object
      properties:
        title:
          type: string
        organization:
          type: string
        type:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        current:
          type: boolean
        location:
          type: string

    Skill:
      type: object
      required: [id, name, category, proficiency, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        proficiency:
          type: integer
        iconName:
          type: ["string", "null"]
        createdAt:
          type: string

    SkillInput:
      type: object
      required: [name, category, proficiency]
      properties:
        name:
          type: string
        category:
          type: string
        proficiency:
          type: integer
        iconName:
          type: string

    SkillUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        proficiency:
          type: integer
        iconName:
          type: string

    Contact:
      type: object
      required: [id, name, email, subject, message, read, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string

    ContactInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string

    AdminCredentials:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean

    StatsSummary:
      type: object
      required: [totalProjects, totalCertificates, totalExperiences, totalSkills, unreadMessages, featuredProjects]
      properties:
        totalProjects:
          type: integer
        totalCertificates:
          type: integer
        totalExperiences:
          type: integer
        totalSkills:
          type: integer
        unreadMessages:
          type: integer
        featuredProjects:
          type: integer

    SkillCategoryCount:
      type: object
      required: [category, count]
      properties:
        category:
          type: string
        count:
          type: integer