# lib/api-zod/src/generated/types/experienceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface ExperienceInput {
  title: string;
  organization: string;
  type: string;
  description?: string;
  startDate: string;
  endDate?: string;
  current?: boolean;
  location?: string;
}