# lib/api-zod/src/generated/types/skillUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface SkillUpdate {
  name?: string;
  category?: string;
  proficiency?: number;
  iconName?: string;
}